+++ spring-integration-core/src/test/java/org/springframework/integration/message/MessageBuilderTests.java
@@ -116,14 +116,14 @@ public class MessageBuilderTests {
 	@Test
 	public void testPriority() {
 		Message<Integer> importantMessage = MessageBuilder.withPayload(1)
-			.setPriority((long)123).build();
+			.setPriority(123).build();
 		assertEquals(new Long(123), importantMessage.getHeaders().getPriority());
 	}
 
 	@Test
 	public void testNonDestructiveSet() {
 		Message<Integer> message1 = MessageBuilder.withPayload(1)
-			.setPriority((long)42).build();
+			.setPriority(42).build();
 		Message<Integer> message2 = MessageBuilder.fromMessage(message1)
 			.setHeaderIfAbsent(MessageHeaders.PRIORITY, 13)
 			.build();
